+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v1.0.1 (8/6/2016)
+
+- Added `Receipt#valid?` method to verify Merkle tree proof.
+- Validate `Receipt` merkle tree and target hash match on import.
+
 ## v1.0.0 (8/5/2016)
 
 - Refactor Ruby API to better match the Client, HashItem, Receipt, Confirmation hierarchy.